backend/requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
llama-index-core>=0.10.0
llama-index-indices-managed-llama-cloud>=0.1.0
openai>=1.12.0
typing-extensions>=4.5.0
python-multipart==0.0.6
pydantic<3.0.0,>=2.5.2
python-jose==3.3.0
requests==2.31.0
uvicorn==0.27.1
asyncio==3.4.3
quart>=0.19.4
quart-cors>=0.7.0
hypercorn==0.15.0
platformdirs>=4.0.0
# Pour une future implémentation d'agents
agno>=0.2.0
duckduckgo-search>=4.0.0
lancedb>=0.4.0
# Fin des nouvelles dépendances


